requirements.txt
fastapi
deta
passlib[bcrypt]
python-jose[cryptography]
